feat: 중개인 매물 등록 기능 추가 [S12P11D205-208]

+++ backend/igeolu/src/main/java/com/ssafy/igeolu/domain/property/repository/PropertyRepository.java
@@ -1,10 +1,13 @@
 package com.ssafy.igeolu.domain.property.repository;
 
+import java.util.List;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import com.ssafy.igeolu.domain.property.entity.Property;
+import com.ssafy.igeolu.domain.user.entity.User;
 
 @Repository
-public interface PropertyRepository extends JpaRepository<Property, Long> {
+public interface PropertyRepository extends JpaRepository<Property, Integer> {
 }
